# openApi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API de Voluntariado Inteligente para Crises",
    "description": "API de orquestração que conecta instituições em situação de crise com voluntários qualificados, utilizando correspondência baseada em habilidades (skill-based volunteering) e proximidade geográfica. Projetada para integração com o IBM watsonx Orchestrate como Skill e com o watsonx.ai (Granite) para extração de habilidades a partir de relatos livres em linguagem natural.",
    "contact": {
      "name": "Equipe Hackathon UNASP 2026"
    },
    "version": "0.2.0"
  },
  "servers": [
  {
    "url": "https://coveting-economy-tingling.ngrok-free.dev",
    "description": "Servidor Local via ngrok"
  }
  ],
  "paths": {
...
    "/api/v1/vinculos": {
      "post": {
        "tags": [
          "Vínculos"
        ],
        "summary": "Executar matchmaking para uma necessidade",
        "description": "Roda o algoritmo de correspondência inteligente e cria registros de proposta de vínculo (status `proposto`) para os top-N voluntários mais aderentes. O score combina skill match, proximidade geográfica, reputação e disponibilidade recente (anti-burnout), com pesos ajustados conforme o nível de urgência da necessidade.",
        "operationId": "executar_matchmaking_api_v1_vinculos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VinculoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchmakingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vinculos/{vinculo_id}": {
      "get": {
        "tags": [
          "Vínculos"
        ],
        "summary": "Consultar vínculo",
        "operationId": "obter_vinculo_api_v1_vinculos__vinculo_id__get",
        "parameters": [
          {
            "name": "vinculo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vinculo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinculoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vinculos/{vinculo_id}/aceitar": {
      "post": {
        "tags": [
          "Vínculos"
        ],
        "summary": "Voluntário aceita o vínculo",
        "description": "Quando aceito, a necessidade transita para `em_atendimento` e o voluntário para `em_missao`. Este endpoint expõe o vínculo completo (inclusive contato) para a instituição consultar.",
        "operationId": "aceitar_vinculo_api_v1_vinculos__vinculo_id__aceitar_post",
        "parameters": [
          {
            "name": "vinculo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vinculo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinculoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vinculos/{vinculo_id}/recusar": {
      "post": {
        "tags": [
          "Vínculos"
        ],
        "summary": "Voluntário recusa o vínculo",
        "operationId": "recusar_vinculo_api_v1_vinculos__vinculo_id__recusar_post",
        "parameters": [
          {
            "name": "vinculo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vinculo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinculoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vinculos/{vinculo_id}/concluir": {
      "post": {
        "tags": [
          "Vínculos"
        ],
        "summary": "Registrar conclusão do atendimento",
        "description": "Marca o vínculo como concluído e libera o voluntário (volta para `disponivel`). A necessidade é marcada como `resolvida`.",
        "operationId": "concluir_vinculo_api_v1_vinculos__vinculo_id__concluir_post",
        "parameters": [
          {
            "name": "vinculo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vinculo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinculoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vinculos/{vinculo_id}/feedback": {
      "post": {
        "tags": [
          "Vínculos"
        ],
        "summary": "Registrar feedback pós-atendimento",
        "description": "Fecha o loop agêntico: o feedback ajusta a reputação do voluntário via média móvel simples, realimentando o ranking em futuros matches.",
        "operationId": "registrar_feedback_api_v1_vinculos__vinculo_id__feedback_post",
        "parameters": [
          {
            "name": "vinculo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vinculo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FeedbackCreate": {
        "properties": {
          "voluntario_compareceu": {
            "type": "boolean",
            "title": "Voluntario Compareceu",
            "example": true
          },
          "skill_adequada": {
            "type": "boolean",
            "title": "Skill Adequada",
            "example": true
          },
          "nota": {
            "type": "integer",
            "maximum": 5,
            "minimum": 1,
            "title": "Nota",
            "example": 5
          },
          "comentario": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comentario"
          }
        },
        "type": "object",
        "required": [
          "voluntario_compareceu",
          "skill_adequada",
          "nota"
        ],
        "title": "FeedbackCreate"
      },
      "FeedbackOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "vinculo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Vinculo Id"
          },
          "voluntario_compareceu": {
            "type": "boolean",
            "title": "Voluntario Compareceu"
          },
          "skill_adequada": {
            "type": "boolean",
            "title": "Skill Adequada"
          },
          "nota": {
            "type": "integer",
            "title": "Nota"
          },
          "comentario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comentario"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vinculo_id",
          "voluntario_compareceu",
          "skill_adequada",
          "nota",
          "comentario",
          "criado_em"
        ],
        "title": "FeedbackOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchRecomendado": {
        "properties": {
          "vinculo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Vinculo Id"
          },
          "voluntario": {
            "$ref": "#/components/schemas/VoluntarioPublico"
          },
          "score_match": {
            "type": "number",
            "title": "Score Match",
            "description": "Score composto entre 0 e 1."
          },
          "habilidades_correspondentes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Habilidades Correspondentes"
          },
          "distancia_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distancia Km"
          },
          "status": {
            "$ref": "#/components/schemas/StatusVinculo"
          }
        },
        "type": "object",
        "required": [
          "vinculo_id",
          "voluntario",
          "score_match",
          "habilidades_correspondentes",
          "distancia_km",
          "status"
        ],
        "title": "MatchRecomendado",
        "description": "Recomendação individual retornada pelo matchmaking."
      },
      "MatchmakingResponse": {
        "properties": {
          "necessidade_id": {
            "type": "string",
            "format": "uuid",
            "title": "Necessidade Id"
          },
          "nivel_urgencia": {
            "$ref": "#/components/schemas/NivelUrgencia"
          },
          "total_recomendados": {
            "type": "integer",
            "title": "Total Recomendados"
          },
          "recomendacoes": {
            "items": {
              "$ref": "#/components/schemas/MatchRecomendado"
            },
            "type": "array",
            "title": "Recomendacoes"
          }
        },
        "type": "object",
        "required": [
          "necessidade_id",
          "nivel_urgencia",
          "total_recomendados",
          "recomendacoes"
        ],
        "title": "MatchmakingResponse"
      },
...
      "NivelUrgencia": {
        "type": "string",
        "enum": [
          "baixo",
          "medio",
          "alto",
          "critico"
        ],
        "title": "NivelUrgencia"
      },
...
      "StatusVinculo": {
        "type": "string",
        "enum": [
          "proposto",
          "aceito",
          "recusado",
          "em_atendimento",
          "concluido",
          "cancelado"
        ],
        "title": "StatusVinculo"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VinculoCreate": {
        "properties": {
          "necessidade_id": {
            "type": "string",
            "format": "uuid",
            "title": "Necessidade Id",
            "description": "ID da necessidade para a qual se deseja encontrar voluntários."
          },
          "top_n": {
            "type": "integer",
            "maximum": 20,
            "minimum": 1,
            "title": "Top N",
            "description": "Quantidade máxima de voluntários a recomendar.",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "necessidade_id"
        ],
        "title": "VinculoCreate",
        "description": "Solicitação de matchmaking para uma necessidade já registrada."
      },
      "VinculoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "necessidade_id": {
            "type": "string",
            "format": "uuid",
            "title": "Necessidade Id"
          },
          "voluntario_id": {
            "type": "string",
            "format": "uuid",
            "title": "Voluntario Id"
          },
          "status": {
            "$ref": "#/components/schemas/StatusVinculo"
          },
          "score_match": {
            "type": "number",
            "title": "Score Match"
          },
          "habilidades_correspondentes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Habilidades Correspondentes"
          },
          "distancia_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distancia Km"
          },
          "proposto_em": {
            "type": "string",
            "format": "date-time",
            "title": "Proposto Em"
          },
          "aceito_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aceito Em"
          },
          "concluido_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concluido Em"
          }
        },
        "type": "object",
        "required": [
          "id",
          "necessidade_id",
          "voluntario_id",
          "status",
          "score_match",
          "habilidades_correspondentes",
          "distancia_km",
          "proposto_em",
          "aceito_em",
          "concluido_em"
        ],
        "title": "VinculoOut",
        "description": "Visão completa de um vínculo — expõe contato após aceite."
      },
...
      "VoluntarioPublico": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "iniciais": {
            "type": "string",
            "title": "Iniciais",
            "description": "Apenas iniciais do nome (ex.: 'A. C. S.')"
          },
          "habilidades": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Habilidades"
          },
          "reputacao": {
            "type": "number",
            "title": "Reputacao"
          }
        },
        "type": "object",
        "required": [
          "id",
          "iniciais",
          "habilidades",
          "reputacao"
        ],
        "title": "VoluntarioPublico",
        "description": "Versão reduzida exposta antes do aceite de vínculo (LGPD)."
      },
...
    }
  },
...
}